+++ Employee-Management-System/src/main/java/com/capg/controller/ProjectController.java
@@ -19,6 +19,7 @@ import com.capg.entity.Manager;
 import com.capg.entity.Project;
 import com.capg.exception.EmployeeNotFoundException;
 import com.capg.exception.EmployeesEmptyException;
+import com.capg.exception.EnterValidDetailsException;
 import com.capg.exception.ProjectEmptyException;
 import com.capg.exception.ProjectNotFoundException;
 import com.capg.service.ManagementService;
@@ -31,25 +32,40 @@ public class ProjectController {
 
 	@PostMapping("/save-project")
 	public ResponseEntity<Project> addProject(@RequestBody Project project) {
+		if(project.getId()<0) {		
+			throw new EnterValidDetailsException("Please Enter Valid Project Id");
+		}
     	return new ResponseEntity<Project>(managementService.saveProject(project), HttpStatus.ACCEPTED);	
 	}
 
 	 @GetMapping("/{projectId}")
 	    public ResponseEntity<Project> getByProjectId(@PathVariable int projectId) {
+		 	if(projectId<0) {
+		 		throw new EnterValidDetailsException("Please Enter Valid Project Id");
+		 	}
+		 	else {
 	    	if(!managementService.findProjectById(projectId).isPresent()) {
 				throw new EmployeeNotFoundException("Project not found with projectId "+ projectId);
 			}
 	    	return new ResponseEntity<Project>(managementService.findProjectById(projectId).get(), HttpStatus.FOUND);	
 	    }
+	 }
 
 	 
 	@PutMapping("/{projectId}/set-manager/{managerId}")
 	private ResponseEntity<Project> assignManagerToProject(@PathVariable int projectId, @PathVariable int managerId) {
+		
+		if(projectId <0 || managerId<0) {
+			
+			throw new EnterValidDetailsException("Either Project Id or Manager Id is Invalid, Please Enter Valid Id's");
+		}
+		else {
 		Project project = managementService.findProjectById(projectId).get();
 		Manager manager = managementService.findManagerById(managerId).get();
 		project.setManager(manager);
     	return new ResponseEntity<Project>(managementService.saveProject(project), HttpStatus.ACCEPTED);	
 	}
+}
 
 	@GetMapping("")
 	public ResponseEntity<List<Project>> getAllProjects() {
@@ -63,10 +79,16 @@ public class ProjectController {
 	
 	@DeleteMapping("/delete-by-id/{projectId}")
     public void deleteByid(@PathVariable int projectId) {
+		
+		if(projectId<0) {
+			throw new EnterValidDetailsException("Please Enter Valid Project Id");	
+		}
+		else {
     	if (!managementService.findProjectById(projectId).isPresent()) {
 			throw new ProjectNotFoundException("Does not found Project with " + projectId);
 		}
     	 managementService.deleteByProjectId(projectId);
     }
+}
 
 }
